+++ src/granularity_convert/CMakeLists.txt
@@ -1,4 +1,7 @@
 target_sources(plateau PRIVATE
         "granularity_converter.cpp"
         "node_path.cpp"
+        "convert_to_atomic.cpp"
+        "convert_from_atomic_to_primary.cpp"
+        "merge_primary_node_and_children.cpp"
 )
